[MOVE] UserJourneyStep to a new user_journey_step module
