+++ efootprint/api_utils/json_to_system.py
@@ -12,7 +12,8 @@ from efootprint.core.hardware.servers.serverless import Serverless
 from efootprint.core.hardware.servers.on_premise import OnPremise
 from efootprint.core.hardware.hardware_base_classes import Hardware
 from efootprint.core.usage.usage_pattern import UsagePattern
-from efootprint.core.usage.user_journey import UserJourney, UserJourneyStep
+from efootprint.core.usage.user_journey import UserJourney
+from efootprint.core.usage.user_journey_step import UserJourneyStep
 from efootprint.core.hardware.network import Network
 from efootprint.core.hardware.device_population import DevicePopulation
 from efootprint.constants.countries import Country
